Create scrollbars.css
This is themed after ZeroNet's fixbutton. Hover over it to see it turn the same shade purple as the fixbutton hovered. This only works in browsers using ```::-webkit``` selectors and a fine pointer such as a mouse*.

*There's a jittering/flickering bug on touch devices like Android, so this CSS excludes them.

Link to it:  ```<link rel="stylesheet" href="/uimedia/scrollbars.css" />```

Would be nice to see this used on more sites such as ZeroHello and ZeroTalk.